+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 2018-11-18
+### Added
+- Logging for the fetch process and reporting merge statistics
+- Creates a `CensusDataStore` class to handle all calls to the Census API.  All calls are logged at the debug level so progress can be seen.  All data fetches will retry again in two minutes (up to 10 times) if there are any issues with the connection of availability of the Census API.
+- Deploys a log file to S3 after demographics are built for a geography
+
+### Changed
+- `create_00_weights.py` has been adjusted so appropriate weights are calculated for counts and rates
+
 ## 2018-11-10
 ### Added
   - changelog